// Font-End/Font-end Admin-3/src/app/shared/models/B20AccDoc_Purchase.ts
export class vB20AccDoc_Purchase {


    id: number;
    stt: string;
    docStatus: number;

    docCode: string;
    docGroup: string;
    docDate: string;
    fiscalYear: string;
    customerCode: string;
    docNo: string;
    trackingNo: string;
    customerName: string;
    address: string;
    email: string;
    phone: string;
    note: string;
    currencyCode: string;
    exchangeRate: number;
    paymentMethod: string;
    paymentStatus: string;
    dueDate: number;
    totalOriginalAmount: number;
    totalAmount: number;
    totalOriginalAmount0: number;
    totalAmount0: number;
    totalOriginalAmount3: number;
    totalAmount3: number;
    totalOriginalAmount4: number;
    totalAmount4: number;
    isActive: boolean;
    createdBy: number;
    createdAt: string;
    modifiedBy: number;
    modifiedAt: string;
    docStatusName: string;
    paymentName: string;
    payStatusName: string;
    taxCode: string;
    warehouseCode: string;
    taxRate: number;
    location: string;
    loctionName: string;

    vB20AccDocPurchase_Json: Array<any>;
}